# bajaj-admin/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medic 24x7 App",
        "description": "REST API's for Medic24x7 application",
        "termsOfService": "",
        "contact": {
            "email": "sayalib@leometric.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "LOGIN CUSTOMER"
                ],
                "summary": "Send OTP/Login",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
        "/api/v1/user/bookmark/list": {
            "get": {
                "tags": [
                    "Bookmark"
                ],
                "summary": "Bookmark List",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/category/list": {
            "get": {
                "tags": [
                    "Category List Partner/User"
                ],
                "summary": "Category List",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/partner/list": {
            "get": {
                "tags": [
                    "Partner List "
                ],
                "summary": "Partner List",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/createAmbulance": {
            "post": {
                "tags": [
                    "Ambulance"
                ],
                "summary": "Insert ambulance",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "partner_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "ambulance_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "total_amb",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "amb_occupied",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "about",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/updateAmbulance/{id}": {
            "put": {
                "tags": [
                    "Ambulance"
                ],
                "summary": "Update ambulance",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "total_amb",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "amb_occupied",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "about",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/createBed": {
            "post": {
                "tags": [
                    "Beds"
                ],
                "summary": "Insert Bed",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "partner_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "total_bed",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "bed_occupied",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/updateClinic/{id}": {
            "put": {
                "tags": [
                    "Clinic"
                ],
                "summary": "Update Clinic",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "min_charges",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/updateDoctor/{id}": {
            "put": {
                "tags": [
                    "Doctor"
                ],
                "summary": "Update ambulance",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "min_charges",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/login": {
            "post": {
                "tags": [
                    "LOGIN PARTNER"
                ],
                "summary": "Send OTP/Login",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/admin/paymethod": {
            "get": {
                "tags": [
                    "Payment Method"
                ],
                "summary": "Payment Method List",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/createRating": {
            "post": {
                "tags": [
                    "Ratings/Review"
                ],
                "summary": "Insert Rating/Review",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "partner_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "customer_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "ratings",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "review",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/state/list": {
            "get": {
                "tags": [
                    "State-City List USER"
                ],
                "summary": "State List",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/city/list": {
            "get": {
                "tags": [
                    "State-City List USER"
                ],
                "summary": "City List",
                "operationId": "city",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "state_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}